feat: 관심 도서 중복 방지 및 취소 기능

+++ bookbook.js
@@ -13,16 +13,33 @@ let totalResults = 0;
 let groupSize = 5;
 let currentPage = 1;
 
-const saveToLocalStorage = (book) => {
+const saveToLocalStorage = (book, button) => {
   let favoriteBooks = JSON.parse(localStorage.getItem("favoriteBooks")) || [];
 
-  // 개수가 넘으면 알림을 띄우는 거 말고 최신순으로 반영하게끔 했습니다
-  if (favoriteBooks.length >= 20) {
-    favoriteBooks.shift();
-  }
+  const bookIndex = favoriteBooks.findIndex(
+    (favoriteBook) => favoriteBook.title === book.title
+  );
 
-  favoriteBooks.push(book);
-  localStorage.setItem("favoriteBooks", JSON.stringify(favoriteBooks));
+  const favoriteImg = button.querySelector("img");
+
+  if (bookIndex !== -1) {
+    const removeConfirmation = confirm("이 책의 즐겨찾기를 해제하시겠습니까?");
+    if (removeConfirmation) {
+      favoriteBooks.splice(bookIndex, 1);
+      localStorage.setItem("favoriteBooks", JSON.stringify(favoriteBooks));
+      favoriteImg.src = "./heart.png";
+    } else {
+      favoriteImg.src = "./fullHeart.png";
+    }
+  } else {
+    if (favoriteBooks.length >= 20) {
+      favoriteBooks.shift();
+    }
+
+    favoriteBooks.push(book);
+    localStorage.setItem("favoriteBooks", JSON.stringify(favoriteBooks));
+    favoriteImg.src = "./fullHeart.png";
+  }
 };
 
 const addFavoriteBtnEvent = () => {
@@ -48,15 +65,7 @@ const addFavoriteBtnEvent = () => {
         publisher,
       };
 
-      saveToLocalStorage(bookData);
-
-      const favoriteImg = button.querySelector("img");
-
-      if (favoriteImg.src.includes("heart.png")) {
-        favoriteImg.src = "./fullHeart.png";
-      } else {
-        favoriteImg.src = "./heart.png";
-      }
+      saveToLocalStorage(bookData, button);
     });
   });
 };
